+++ Cargo.toml
@@ -17,6 +17,7 @@ convert_case = "0.5.0"
 directories = "4.0.1"
 futures-util = "0.3.21"
 indicatif = "0.16.2"
+lazy_static = "1.4.0"
 regex = "1.5.6"
 reqwest = { version = "0.11.11", features = ["stream"] }
 ron = "0.7.0"
