+++ src/api/mod.rs
@@ -1,6 +1,9 @@
-use crate::model::Mod;
 use reqwest::Client;
 use serde_json::Value;
+
+pub mod model;
+
+use model::Mod;
 
 pub async fn get_package_index() -> Result<Vec<Mod>, String> {
     let client = Client::new();
